In class Mainmenu.h, resolve norrowing conversion

+++ GUI/MainMenu.cpp
@@ -105,7 +105,8 @@ void MainMenu::render() {
 }
 
 void MainMenu::update() {
-    backGround.setTextureRect(sf::IntRect(0, 0, renderWin->getView().getSize().x, renderWin->getView().getSize().y));
+    backGround.setTextureRect(sf::IntRect(0, 0, static_cast<int>(renderWin->getView().getSize().x),
+                                          static_cast<int>(renderWin->getView().getSize().y)));
     auto mousePos = sf::Mouse::getPosition(*renderWin);
     auto worldPos = renderWin->mapPixelToCoords(mousePos);
     if (!nextLevel) {
